Move test to package
