+++ src/test/scala/biowdl/tests/reassembly/ReAssemblySuccess.scala
@@ -19,26 +19,18 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+package biowdl.tests.reassembly
+
 import java.io.File
 
-import nl.biopet.utils.biowdl.Pipeline
+import nl.biopet.utils.biowdl.PipelineSuccess
 
-trait ReAssembly extends Pipeline {
-  def inputAssembly: File
+trait ReAssemblySuccess extends ReAssembly with PipelineSuccess {
+  val unmapppedReadsFilteredFile: File = createFile("unmapppedReadsFiltered.bam")
+  val readsMappedToInputAssemblyFile: File = createFile("ReadsMappedToInputAssembly.bam")
+  val filteredR1File: File = createFile("filtered_reads1.fq.gz")
+  val filteredR2File: Option[File] = createOptionalFile(read2.isDefined, "filtered_reads2.fq.gz")
 
-  def read1: File
-
-  def read2: Option[File]
-
-  def startFile: File = new File("ReAssembly.wdl")
-
-  override def inputs: Map[String, Any] =
-    super.inputs ++
-      Map(
-        "ReAssembly.read1" -> read1.getAbsolutePath,
-        "ReAssembly.inputAssembly" -> inputAssembly.getAbsolutePath,
-        "ReAssembly.outputDir" -> outputDir.getAbsolutePath
-      ) ++
-      read2.map("ReAssembly.read2" -> _.getAbsolutePath)
-
+  val scaffoldsFastaFile: File = createFile("spades", "scaffolds.fasta")
+  val contigsFastFile: File = createFile("spades", "contigs.fasta")
 }
